requirements.txt
openai
python-dotenv
langchain-nvidia-ai-endpoints
langchain
langchain_community 
langchain_core
faiss-cpu
streamlit
pypdf
